<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->
Base on  #5748 and #5638 , for flashinfer mla, remove q and k cat.

### Accuracy
```
Accuracy: 0.951
Invalid: 0.000
Latency: 228.672 s
Output throughput: 554.173 token/s
```

### Performance
main branch:
```
{"run_name": "default", "batch_size": 1, "input_len": 1024, "output_len": 1024, "latency": 14.4687, "output_throughput": 70.77, "overall_throughput": 141.55}

{"run_name": "default", "batch_size": 16, "input_len": 1024, "output_len": 1024, "latency": 28.8723, "output_throughput": 567.47, "overall_throughput": 1134.93}

{"run_name": "default", "batch_size": 32, "input_len": 1024, "output_len": 1024, "latency": 38.0349, "output_throughput": 861.52, "overall_throughput": 1723.05}
```

this PR:
```
{"run_name": "default", "batch_size": 1, "input_len": 1024, "output_len": 1024, "latency": 14.5066, "output_throughput": 70.59, "overall_throughput": 141.18}

{"run_name": "default", "batch_size": 16, "input_len": 1024, "output_len": 1024, "latency": 28.4372, "output_throughput": 576.15, "overall_throughput": 1152.29}

{"run_name": "default", "batch_size": 32, "input_len": 1024, "output_len": 1024, "latency": 37.2972, "output_throughput": 878.57, "overall_throughput": 1757.13}
```

### Profile

#### Prefill
main branch:
![image](https://github.com/user-attachments/assets/b5e21f35-b5da-4c03-8961-a24afdf36402)

this PR:
![image](https://github.com/user-attachments/assets/c916a15b-0236-44a0-8330-1ce8575d0e34)

47 us to 3us

#### Decode 
main branch bs=1 cuda graph+torch compile:
![image](https://github.com/user-attachments/assets/cb82a756-e01b-4324-b8dd-48d2ba09de52)

this PR bs=1 cuda graph+torch compile:
![image](https://github.com/user-attachments/assets/c80eb6e2-03a1-4294-8429-7297b44739bf)
bs=1, cuda graph+torch compile, almost the same

main branch bs=1 cuda graph+ without torch compile fused with other ops:
![image](https://github.com/user-attachments/assets/6999123c-faad-49a4-a8e5-5302ae7e156c)

this PR bs=1 cuda graph+ without torch compile fused with other ops:
![image](https://github.com/user-attachments/assets/9b27b16c-058f-4f76-8c3e-e7bba0645067)
bs=1, cuda graph, without torch compile, 6~7 us -> 1us

## Modifications

<!-- Describe the changes made in this PR. -->
- Update deepseek_v2 code, remove q and k cat.
- In flashinfer_mla_backend: 
> Cat when ragged, no cat in other scenes. 
> Use set_mla_kv_buffer when k_rope is not empty

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.